% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreQuran-package.R
\docType{package}
\name{exploreQuran-package}
\alias{exploreQuran}
\alias{exploreQuran-package}
\title{exploreQuran: Quran Translation Analytic}
\description{
Text mining and analytic for Quran translation from tanzil.net
}
\keyword{internal}
