% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_surah.r
\name{tf_trans}
\alias{tf_trans}
\title{Create term frequency by surah}
\usage{
tf_trans(tanzil_trans_object, surah_number = (1:114))
}
\arguments{
\item{tanzil_trans_object}{Tanzil translation object created with `tanzil_translation()` function}

\item{surah_number}{Surah number. Default to all 1-114 surah. Can refer to `quran_index` for surah number and Quran index.}
}
\value{
A term frequency data frame for selected surah
}
\description{
Create term frequency by surah
}
\examples{

trans_en_sahih <- tanzil_translation("https://tanzil.net/trans/en.sahih")
trans_en_yusof <- tanzil_translation("https://tanzil.net/trans/en.yusufali")


tf_trans_shahih <- tf_trans(trans_en_sahih)
tf_trans_surah_short_shahih <- tf_trans(trans_en_sahih, c(113, 112, 109, 111, 114, 110))
tf_trans_surah_short_yusoff <- tf_trans(trans_en_yusof, c(113, 112, 109, 111, 114, 110))




}
