% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cloud_surah.r
\name{wordcloud_trans}
\alias{wordcloud_trans}
\title{Create Quran translation's word cloud}
\usage{
wordcloud_trans(
  tanzil_trans_object,
  surah_number = (1:114),
  max_word = 200,
  min_freq = 1
)
}
\arguments{
\item{tanzil_trans_object}{Tanzil translation object created with `tanzil_translation()` function}

\item{surah_number}{Surah number. Default to all 1-114 surah. Can refer to `quran_index` for surah number and Quran index.}

\item{max_word}{Maximum number of words to be plotted. least frequent terms dropped}

\item{min_freq}{words with frequency below min.freq will not be plotted}
}
\value{
Plot of word cloud
}
\description{
Create Quran translation's word cloud
}
\examples{

trans_en_sahih <- tanzil_translation("https://tanzil.net/trans/en.sahih")
wordcloud_trans(trans_en_sahih, c(1,2,3))


}
